replaced double with float

+++ JEE/Java/Employe.java
@@ -1,6 +1,6 @@
 public class Employe{
     private String nom, prenom;
-    private double salaire;
+    private float salaire;
 
     // Constructeurs
     public Employe() {}
@@ -10,7 +10,7 @@ public class Employe{
         prenom = p;
     }
 
-    public Employe(String n, String p, double s){
+    public Employe(String n, String p, float s){
         nom = n;
         prenom = p;
         salaire = s;
@@ -32,11 +32,11 @@ public class Employe{
         this.prenom = prenom;
     }
  
-    public double getSalaire() {
+    public float getSalaire() {
         return Salaire;
     }
  
-    public void setSalaire(double Salaire) {
+    public void setSalaire(float Salaire) {
         this.Salaire = Salaire;
     }
